ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic-settings
motor
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
ultralytics
opencv-python
